Split Elevators and DamageStrictness from layout_configuration
