+++ test/interface_common/test_preset_editor.py
@@ -4,8 +4,8 @@ from unittest.mock import MagicMock
 import pytest
 
 from randovania.interface_common.preset_editor import PresetEditor
-from randovania.layout.layout_configuration import LayoutElevators, \
-    LayoutSkyTempleKeyMode
+from randovania.layout.layout_configuration import LayoutSkyTempleKeyMode
+from randovania.layout.elevators import LayoutElevators
 
 
 @pytest.fixture(name="editor")
