+++ test/layout/test_layout_configuration.py
@@ -10,8 +10,9 @@ from randovania.layout.ammo_configuration import AmmoConfiguration
 from randovania.layout.available_locations import AvailableLocationsConfiguration
 from randovania.layout.beam_configuration import BeamConfiguration
 from randovania.layout.hint_configuration import HintConfiguration
-from randovania.layout.layout_configuration import LayoutConfiguration, LayoutSkyTempleKeyMode, \
-    LayoutElevators, LayoutDamageStrictness
+from randovania.layout.layout_configuration import LayoutConfiguration, LayoutSkyTempleKeyMode
+from randovania.layout.elevators import LayoutElevators
+from randovania.layout.damage_strictness import LayoutDamageStrictness
 from randovania.layout.major_items_configuration import MajorItemsConfiguration
 from randovania.layout.starting_location import StartingLocation
 from randovania.layout.translator_configuration import TranslatorConfiguration
